+++ tests/integration/cases/call_zip_dict_values/JavaScript_call@es2015.js
@@ -1,0 +1,4 @@
+function process() {}
+function emit() {}
+emit(process({ value: "hello" }), {"a": 1, "b": 2});
+emit(process({ value: 42 }), {"c": 3, "d": 4});