+++ tests/integration/cases/call_zip_dict_values/Perl_call@v5_36.pl
@@ -1,0 +1,4 @@
+sub process {}
+sub emit {}
+emit(process("hello"), {"a" => 1, "b" => 2});
+emit(process(42), {"c" => 3, "d" => 4});