+++ tests/integration/cases/call_zip_dict_values/PowerShell_call@v7.ps1
@@ -1,0 +1,4 @@
+function process {}
+function emit {}
+emit(process("hello"), @{"a" = 1; "b" = 2})
+emit(process(42), @{"c" = 3; "d" = 4})